requirements.txt
newsapi-python==0.2.6
